Criação dos endpoints dos projetos

+++ api/src/main/kotlin/com/api/application/project/exceptions/ProjectTypeExceptions.kt
@@ -1,13 +1,13 @@
 package com.api.application.project.exceptions
 
 sealed class ProjectTypeExceptions(message: String): Exception(message) {
-    abstract val projectTypeID: Int?
+    abstract val projectTypeID: Long?
 }
 
 data class ProjectTypeNotFoundException (
-    override val projectTypeID: Int?
+    override val projectTypeID: Long?
 ): ProjectTypeExceptions("Tipo de projeto $projectTypeID não encontrado!")
 
 data class ProjectTypeNotInsertException (
-    override val projectTypeID: Int?
+    override val projectTypeID: Long?
 ): ProjectTypeExceptions("Não foi possivel inserir o tipo de projeto de ID: $projectTypeID!")
